+++ src/environments/environment.prod.example.ts
@@ -1,0 +1,12 @@
+export const environment = {
+  production: true,
+  firebaseConfig: {
+    apiKey: '',
+    authDomain: '',
+    databaseURL: '',
+    projectId: '',
+    storageBucket: '',
+    messagingSenderId: '',
+    appId: '',
+  },
+};